Cleaned up organization significantly
The previous iteration worked, but not in the correct way: the linker script was
actually being bypassed, which was why I was required to use the label '_start' in the
startup.s file instead of what the tutorial showed (the linker must have used some sort
of 'sane defaults' which expected that label name).

This version fixes that issue, but also make the improvement that it doesn't rely on any
changes to the cross-compiler directory. The previous commit required that you put
soft-links in a specific folder to 'adalib' and 'adainclude' in the zfp_rts/build/ path.