WIP: make project more agnostic to build environment
edit the run-time project from relying on a modified GPRBuild knowledge
base file to a project-specific set of build parameters (a .cgpr file).
This should make the project more portable as well as less reliant
on user-edited tools

Because the number of flags is quite extensive, first I found and
edited the knowledge base files, and used them to generate a .cgpr file,
which I have checked in for future use.