Eliminate absolute path use in .cgpr
It took quite a bit of fighting and research, but I was able to
eliminate absolute paths in the .cgpr file of the Run-Time sub-project.

I had to manually massage the .cgpr file to use the `--RTS` and
`Project'Project_Dir` directives. There MIGHT be a better way, but I
have yet to find it in the docs.

I also renamed the project file/name and removed some incomplete sources
which seemed to get more things working as intended!