Consolidate 'zfp_rts/src/' into 'zfp_rts/build/adainclude/' folder
I thought long and hard about this, because the impression I got for
WHY the Ada Bare Bones author originally had a "src" folder in the
run-time tutorial was that it could include things that were platform-
agnostic, while the "adainclude" had platform-specific things.
However, it ultimately didn't make sense for my use-case.

My plan is to have a separate "zfp_rts" folder for each target in the
short-term, because honestly the way GCC/GPRConfig assumes a specific
folder structure basically forces me to do things that way. I can
separate things out for different targets in a better way in the future